refactor: resolve #295 - split testCasesData.ts from 540 to under 500 lines
Split into:
- testCasesNotes.ts (194 lines) — Basic Notes, Octaves, Sharp Notes
- testCasesTiming.ts (157 lines) — Tempo, Length Codes, Rests
- testCasesTimbre.ts (205 lines) — Duty Cycle, Envelope, Volume, Multi-Channel, Complete Songs

Closes #295
